Lesson3. continue working

+++ src/main/java/ru/geekbrains/spring/application/core/Product.java
@@ -3,9 +3,9 @@ package ru.geekbrains.spring.application.core;
 public class Product {
     private int id;
     private String name;
-    private double cost;
+    private float cost;
 
-    public Product(int id, String name, double cost) {
+    public Product(int id, String name, float cost) {
         this.id = id;
         this.name = name;
         this.cost = cost;
